# backend/static/masterblog_swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Master Blog API",
    "version": "1.0.0",
    "description": "Simple Flask blog posts API"
  },
  "servers": [
    {
      "url": "http://localhost:5002"
    }
  ],
  "tags": [
    {
      "name": "Blogpost's",
      "description": "Blogpost's API - CRUD"
    }
  ],
  "paths": {
    "/api/posts": {
      "get": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Get all blog posts, with optional sorting.",
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by (e.g., author, title).",
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "description": "Sorting direction (asc or desc).",
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "A list of blog posts.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BlogPost"
              }
            }
          },
          "400": {
            "description": "Invalid sorting parameters."
          }
        }
      },
      "post": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Add a new blog post.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Blog post object to be added.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "author": {
                  "type": "string",
                  "description": "Author of the blog post."
                },
                "title": {
                  "type": "string",
                  "description": "Title of the blog post."
                },
                "content": {
                  "type": "string",
                  "description": "Content of the blog post."
                }
              },
              "required": [
                "author",
                "title",
                "content"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created blog post.",
            "schema": {
              "$ref": "#/components/schemas/BlogPost"
            }
          },
          "400": {
            "description": "Invalid input data."
          }
        }
      }
    },
    "/api/posts/search": {
      "get": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Search blog posts based on query parameters.",
        "parameters": [
          {
            "name": "author",
            "in": "query",
            "type": "string",
            "description": "Search for posts by author."
          },
          {
            "name": "title",
            "in": "query",
            "type": "string",
            "description": "Search for posts by title."
          },
          {
            "name": "content",
            "in": "query",
            "type": "string",
            "description": "Search for posts by content."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of blog posts matching the search criteria.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/BlogPost"
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}": {
      "put": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Update an existing blog post.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "ID of the blog post to update.",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Blog post object with updated information.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "author": {
                  "type": "string",
                  "description": "Updated author of the blog post."
                },
                "title": {
                  "type": "string",
                  "description": "Updated title of the blog post."
                },
                "content": {
                  "type": "string",
                  "description": "Updated content of the blog post."
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated blog post.",
            "schema": {
              "$ref": "#/components/schemas/BlogPost"
            }
          },
          "404": {
            "description": "Blog post not found."
          }
        }
      },
      "delete": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Delete a blog post by ID.",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "ID of the blog post to delete.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted blog post.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid post ID."
          },
          "404": {
            "description": "Blog post not found."
          }
        }
      }
    },
    "/api/posts/{post_id}/like": {
      "patch": {
        "tags": [
          "Blogpost's"
        ],
        "summary": "Increment the like count of a specific blog post.",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "description": "ID of the blog post to like.",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "The updated blog post with incremented likes.",
            "schema": {
              "$ref": "#/components/schemas/BlogPost"
            }
          },
          "404": {
            "description": "Blog post not found."
          },
          "400": {
            "description": "Invalid post ID."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "author": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "likes": {
            "type": "integer",
            "format": "int32",
            "description": "Number of likes for the blog post."
          }
        }
      }
    }
  }
}